refactor: navbar explore component
